pyproject.toml
[project]
name = "brotherql-cli"
version = "1.0.0"
description = "CLI for interfacing with brother-ql printer"
authors = [
    {name = "Jamie Levitt",email = "jamielevitt1207@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "zeroconf (>=0.147.0,<0.148.0)",
    "typer (>=0.16.0,<0.17.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "brother-ql (>=0.9.4,<0.10.0)",
    "tabulate (>=0.9.0,<0.10.0)",
]

[tool.poetry]
packages = [{include = "brotherql_cli"}]

[tool.poetry.scripts]
brotherql-cli = "brotherql_cli.cli.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"